Refactored models into a package
