+++ service/models/account.py
@@ -1,0 +1,86 @@
+"""
+Persistent Base class for database CRUD functions
+"""
+
+import logging
+from datetime import date
+from .persistent_base import db, PersistentBase, DataValidationError
+from .address import Address
+
+logger = logging.getLogger("flask.app")
+
+
+######################################################################
+#  A C C O U N T   M O D E L
+######################################################################
+class Account(db.Model, PersistentBase):
+    """
+    Class that represents an Account
+    """
+
+    # Table Schema
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(64))
+    email = db.Column(db.String(64))
+    phone_number = db.Column(db.String(32), nullable=True)  # phone number is optional
+    date_joined = db.Column(db.Date(), nullable=False, default=date.today())
+    addresses = db.relationship("Address", backref="account", passive_deletes=True)
+
+    def __repr__(self):
+        return f"<Account {self.name} id=[{self.id}]>"
+
+    def serialize(self):
+        """Converts an Account into a dictionary"""
+        account = {
+            "id": self.id,
+            "name": self.name,
+            "email": self.email,
+            "phone_number": self.phone_number,
+            "date_joined": self.date_joined.isoformat(),
+            "addresses": [],
+        }
+        for address in self.addresses:
+            account["addresses"].append(address.serialize())
+        return account
+
+    def deserialize(self, data):
+        """
+        Populates an Account from a dictionary
+
+        Args:
+            data (dict): A dictionary containing the resource data
+        """
+        try:
+            self.name = data["name"]
+            self.email = data["email"]
+            self.phone_number = data.get("phone_number")
+            self.date_joined = date.fromisoformat(data["date_joined"])
+            # handle inner list of addresses
+            address_list = data.get("addresses")
+            for json_address in address_list:
+                address = Address()
+                address.deserialize(json_address)
+                self.addresses.append(address)
+        except AttributeError as error:
+            raise DataValidationError("Invalid attribute: " + error.args[0]) from error
+        except KeyError as error:
+            raise DataValidationError(
+                "Invalid Account: missing " + error.args[0]
+            ) from error
+        except TypeError as error:
+            raise DataValidationError(
+                "Invalid Account: body of request contained bad or no data "
+                + str(error)
+            ) from error
+
+        return self
+
+    @classmethod
+    def find_by_name(cls, name):
+        """Returns all Accounts with the given name
+
+        Args:
+            name (string): the name of the Accounts you want to match
+        """
+        logger.info("Processing name query for %s ...", name)
+        return cls.query.filter(cls.name == name)
